requirements.txt
# requirements.txt
botbuilder-core
botbuilder-schema
botbuilder-integration-aiohttp
aiohttp
python-dotenv
requests
pyodbc
